requirements.txt
mysql-connector-python>=8.0.0
requests>=2.28.0
python-dotenv>=1.0.0
dataclasses>=0.6; python_version < '3.7'
numpy>=1.22.0
google-genai>=0.3.0
